requirements/prod.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/prod.in
#
aiofiles==23.2.1
    # via gradio
aiohttp==3.9.1
    # via gradio
aiosignal==1.3.1
    # via aiohttp
altair==5.2.0
    # via gradio
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   starlette
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
boto3==1.34.18
    # via
    #   boto3-extensions
    #   smart-open
boto3-extensions==0.23.0
    # via gantry
botocore==1.34.18
    # via
    #   boto3
    #   boto3-extensions
    #   s3transfer
cachetools==4.2.4
    # via gantry
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   gantry
    #   uvicorn
click-spinner==0.1.10
    # via gantry
colorama==0.4.6
    # via
    #   gantry
    #   halo
    #   log-symbols
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dateparser==1.2.0
    # via gantry
fastapi==0.109.0
    # via gradio
ffmpy==0.3.1
    # via gradio
filelock==3.13.1
    # via
    #   huggingface-hub
    #   torch
fonttools==4.47.2
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.12.2
    # via
    #   gradio-client
    #   huggingface-hub
    #   torch
gantry==0.4.9
    # via -r requirements/prod.in
gradio==3.40.1
    # via -r requirements/prod.in
gradio-client==0.8.0
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h5py==3.10.0
    # via -r requirements/prod.in
halo==0.0.31
    # via gantry
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via
    #   gradio
    #   gradio-client
huggingface-hub==0.20.2
    # via
    #   gradio
    #   gradio-client
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==7.0.1
    # via -r requirements/prod.in
importlib-resources==6.1.1
    # via gradio
isodate==0.6.1
    # via gantry
jinja2==3.1.3
    # via
    #   -r requirements/prod.in
    #   altair
    #   gradio
    #   torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.20.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
linkify-it-py==2.0.2
    # via markdown-it-py
log-symbols==0.0.14
    # via halo
markdown-it-py[linkify]==2.2.0
    # via
    #   gradio
    #   mdit-py-plugins
markupsafe==2.1.3
    # via
    #   gradio
    #   jinja2
marshmallow==3.20.2
    # via
    #   gantry
    #   marshmallow-oneofschema
marshmallow-oneofschema==3.0.1
    # via gantry
matplotlib==3.8.2
    # via gradio
mdit-py-plugins==0.3.3
    # via gradio
mdurl==0.1.2
    # via markdown-it-py
monotonic==1.6
    # via gantry
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
networkx==3.2.1
    # via torch
numpy==1.26.3
    # via
    #   -r requirements/prod.in
    #   altair
    #   contourpy
    #   gantry
    #   gradio
    #   h5py
    #   matplotlib
    #   pandas
    #   torchvision
orjson==3.9.10
    # via gradio
packaging==23.2
    # via
    #   altair
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   marshmallow
    #   matplotlib
pandas==2.1.4
    # via
    #   altair
    #   gantry
    #   gradio
pillow==9.4.0
    # via
    #   -r requirements/prod.in
    #   gradio
    #   matplotlib
    #   torchvision
pydantic==2.5.3
    # via
    #   fastapi
    #   gradio
pydantic-core==2.14.6
    # via pydantic
pydub==0.25.1
    # via gradio
pyngrok==6.0.0
    # via -r requirements/prod.in
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   botocore
    #   dateparser
    #   gantry
    #   matplotlib
    #   pandas
python-multipart==0.0.6
    # via gradio
pytz==2023.3.post1
    # via
    #   dateparser
    #   pandas
pyyaml==6.0.1
    # via
    #   gantry
    #   gradio
    #   huggingface-hub
    #   pyngrok
referencing==0.32.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.12.25
    # via dateparser
requests==2.31.0
    # via
    #   -r requirements/prod.in
    #   gantry
    #   gradio
    #   huggingface-hub
    #   torchvision
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.0
    # via boto3
semantic-version==2.10.0
    # via gradio
six==1.16.0
    # via
    #   halo
    #   isodate
    #   python-dateutil
smart-open[s3]==6.4.0
    # via -r requirements/prod.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
spinners==0.0.24
    # via halo
starlette==0.35.1
    # via fastapi
sympy==1.12
    # via torch
tabulate==0.9.0
    # via gantry
termcolor==2.4.0
    # via halo
toolz==0.12.0
    # via altair
torch==2.1.2
    # via
    #   -r requirements/prod.in
    #   torchvision
torchvision==0.16.2
    # via -r requirements/prod.in
tqdm==4.66.1
    # via
    #   -r requirements/prod.in
    #   gantry
    #   huggingface-hub
typeguard==2.13.3
    # via gantry
typing-extensions==4.9.0
    # via
    #   fastapi
    #   gantry
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
tzdata==2023.4
    # via pandas
tzlocal==5.2
    # via dateparser
uc-micro-py==1.0.2
    # via linkify-it-py
urllib3==2.0.7
    # via
    #   botocore
    #   requests
uvicorn==0.25.0
    # via gradio
websockets==11.0.3
    # via
    #   gradio
    #   gradio-client
yarl==1.9.4
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata
